rename: 파일 이름 케밥 형식으로 변경
