+++ docs/api_exchange.md
@@ -1,11 +1,12 @@
 ## POST /exchange/:exchangedId/accept
 
 ### req template
+
 - description: 포토카드 교환 승인
 - path: /exchange/:exchangedId/accept
 - method: POST
 - params
-    - exchangedId: 교환ID
+  - exchangedId: 교환ID
 
 ### req example
 
@@ -15,17 +16,17 @@
 ### res template
 
 - data
-    - successStatus: 성공 여부
-    - exchangeId: 교환 아이디
-    - shopId: 상점 아이디
-    - description: 교환 설명
-    - id: 교환 시 제시한 카드 아이디
-    - image: 교환 시 제시한 카드 이미지
-    - name: 교환 시 제시한 카드 이름
-    - grade: 교환 시 제시한 카드 등급
-    - genre: 교환 시 제시한 카드 장르
-    - price: 교환 시 제시한 카드 원가
-    - nickname: 교환 시 제시한 카드 원작자
+  - successStatus: 성공 여부
+  - exchangeId: 교환 아이디
+  - shopId: 상점 아이디
+  - description: 교환 설명
+  - id: 교환 시 제시한 카드 아이디
+  - image: 교환 시 제시한 카드 이미지
+  - name: 교환 시 제시한 카드 이름
+  - grade: 교환 시 제시한 카드 등급
+  - genre: 교환 시 제시한 카드 장르
+  - price: 교환 시 제시한 카드 원가
+  - nickname: 교환 시 제시한 카드 원작자
 
 ### res example
 
@@ -41,16 +42,17 @@
   "genre": 0,
   "price": 4,
   "nickname": "코드잇01"
-}
+  }
 
 ## POST /exchange/:exchangedId/refuse
 
 ### req template
+
 - description: 포토카드 교환 거절
 - path: /exchange/:exchangedId/refuse
 - method: POST
 - params
-    - exchangedId: 교환ID
+  - exchangedId: 교환ID
 
 ### req example
 
@@ -60,17 +62,17 @@
 ### res template
 
 - data
-    - successStatus: 성공 여부
-    - exchangeId: 교환 아이디
-    - shopId: 상점 아이디
-    - description: 교환 설명
-    - id: 교환 시 제시한 카드 아이디
-    - image: 교환 시 제시한 카드 이미지
-    - name: 교환 시 제시한 카드 이름
-    - grade: 교환 시 제시한 카드 등급
-    - genre: 교환 시 제시한 카드 장르
-    - price: 교환 시 제시한 카드 원가
-    - nickname: 교환 시 제시한 카드 원작자
+  - successStatus: 성공 여부
+  - exchangeId: 교환 아이디
+  - shopId: 상점 아이디
+  - description: 교환 설명
+  - id: 교환 시 제시한 카드 아이디
+  - image: 교환 시 제시한 카드 이미지
+  - name: 교환 시 제시한 카드 이름
+  - grade: 교환 시 제시한 카드 등급
+  - genre: 교환 시 제시한 카드 장르
+  - price: 교환 시 제시한 카드 원가
+  - nickname: 교환 시 제시한 카드 원작자
 
 ### res example
 
@@ -86,16 +88,17 @@
   "genre": 0,
   "price": 4,
   "nickname": "코드잇01"
-}
+  }
 
 ## DELETE /exchange/:exchangeId
 
 ### req template
+
 - description: 포토 카드 교환 취소
 - path: /exchange/:exchangeId
 - method: DELETE
 - params
-    - exchangeId: 교환ID
+  - exchangeId: 교환ID
 
 ### req example
 
@@ -105,17 +108,17 @@
 ### res template
 
 - data
-    - successStatus: 성공 여부
-    - exchangeId: 교환 아이디
-    - shopId: 상점 아이디
-    - description: 교환 설명
-    - id: 교환 시 제시한 카드 아이디
-    - image: 교환 시 제시한 카드 이미지
-    - name: 교환 시 제시한 카드 이름
-    - grade: 교환 시 제시한 카드 등급
-    - genre: 교환 시 제시한 카드 장르
-    - price: 교환 시 제시한 카드 원가
-    - nickname: 교환 시 제시한 카드 원작자
+  - successStatus: 성공 여부
+  - exchangeId: 교환 아이디
+  - shopId: 상점 아이디
+  - description: 교환 설명
+  - id: 교환 시 제시한 카드 아이디
+  - image: 교환 시 제시한 카드 이미지
+  - name: 교환 시 제시한 카드 이름
+  - grade: 교환 시 제시한 카드 등급
+  - genre: 교환 시 제시한 카드 장르
+  - price: 교환 시 제시한 카드 원가
+  - nickname: 교환 시 제시한 카드 원작자
 
 ### res example
 
@@ -131,4 +134,4 @@
   "genre": 0,
   "price": 4,
   "nickname": "코드잇01"
-}
+  }
