+++ src/middlewares/auth.js
@@ -1,6 +1,6 @@
 import { createCustomError } from "../lib/custom-error.js";
-import ownRepository from "../repositories/ownRepository.js";
-import shopRepository from "../repositories/shopRepository.js";
+import ownRepository from "../repositories/own-repository.js";
+import shopRepository from "../repositories/shop-repository.js";
 import notificationRepository from "../repositories/notification-repository.js";
 import { ownSelect } from "../repositories/selects/own-select.js";
