+++ src/middlewares/validateData.js
@@ -1,14 +1,14 @@
 import { assert } from "superstruct";
-import ownService from "../services/ownService.js";
+import ownService from "../services/own-service.js";
 import { createShopStruct, updateShopStruct } from "../structs/shopStruct.js";
 import { SignUpUser, SignInUser } from "../structs/user-struct.js";
-import shopService from "../services/shopService.js";
+import shopService from "../services/shop-service.js";
 import userService from "../services/user-service.js";
-import ownRepository from "../repositories/ownRepository.js";
+import ownRepository from "../repositories/own-repository.js";
 import exchangeRepository from "../repositories/exchange-repository.js";
-import shopRepository from "../repositories/shopRepository.js";
+import shopRepository from "../repositories/shop-repository.js";
 import { exchangeCardShopAndUserSelect } from "../repositories/selects/exchange-select.js";
-import { shopDetailSelect } from "../repositories/selects/shopSelect.js";
+import { shopDetailSelect } from "../services/selects/shopSelect.js";
 
 export async function validateCreateShopData(req, res, next) {
   const userId = req.session.userId;
@@ -280,7 +280,6 @@ export async function validateExchangeConditions(req, res, next) {
   next();
 }
 
-
 export async function validateExchangeCreator(req, res, next) {
   const { exchangeId } = req.params;
   const userId = req.session.userId;
@@ -338,5 +337,5 @@ export async function checkShopCreator(req, res, next) {
 
   req.body.shopData = isOwner;
 
-  next()
+  next();
 }
