+++ src/repositories/shop-repository.js
@@ -5,7 +5,7 @@ import {
   shopDetailSelect,
   shopListSelect,
   shopOwnerSelect,
-} from "./selects/shopSelect.js";
+} from "../services/selects/shopSelect.js";
 
 async function createShop(createData) {
   return await prisma.shop.create({
