+++ src/services/selects/exchange-select.js
@@ -1,4 +1,4 @@
-import { cardSelect } from "./card-select.js";
+import { cardSelect } from "../../services/selects/card-select.js";
 
 export const exchangeSelect = {
   id: true,
